Problem related to UOM in MRP. https://sourceforge.net/tracker/index.php?func=detail&aid=2605509&group_id=176962&atid=934929
Hi Teo!

I reverted your revision: 8483, the reason is because your approach is not correct.

MRP generate the demand and supply based in UOM define in product , this UOM stocked is use for inventory transaction.

before the change all the MRP qty use QtyOrdered that is the UOM stocked with your change the MRP qty is convert the Qty Entry but it is incorrect.

Business Case:

1.- Define a Product with kg UOM stocked.
2.- Define a UOM Conversion for this product Bags of 40 kg
3.- Create a Sales Order with this product and sell 10 Bags 
4.- MRP use the Qty Ordered to generate the Demand in this case !0 * 40 = 400 kg

With your revision 8483 the Qty Order is convert to Qty Entry in this case 10, but the real demand is 400 kg and no 10 kg

anyway I will like understand what is the Cristi´s issue 

kind regards
Victor Pérez
www.e-evolution.com

git-svn-id: https://adempiere.svn.sourceforge.net/svnroot/adempiere/branches/libero@11258 f8025ce2-341d-0410-ac50-9b82273f5115